feat(plugins): initrd: improve efi img option check
This logic should cover the case where a user supplies a key/cert (in
which case, they should specify creating an EFI image and may have merely
forgotten to set it) while still preserving snakeoil signing an image
that may get made (whether or not to do this is handled by the plugin's
initrd_build script.)

The case which isn't covered is when a user wants a snakeoil signed EFI
image but didn't specify creating an EFI image. It may be useful to raise
a warning here.

Signed-off-by: Dilyn Corner <dilyn.corner@canonical.com>