feat(plugins): initrd: handle key and cert with install_extra
Finding the key and cert is now handled similarly to how finding addons
and firmware are. We've also taken the opportunity to consolidate the
function's handling of source and dest locations, reducing an if/else
into a simple check to which avoids a name nesting problem.

This change now means that all objects supplied via plugin parameters
must be located in $CRAFT_STAGE.

Signed-off-by: Dilyn Corner <dilyn.corner@canonical.com>