feat(plugins): initrd: modify install_extra
Previous iterations would potentially copy the incorrect file from
CRAFT_STAGE to INITRD_ROOT if two files in CRAFT_STAGE had the same name.
Now we at least check the directory we would anticipate the file to be
located in to avoid this.

There may be further issues, such as handling '*'.

Signed-off-by: Dilyn Corner <dilyn.corner@canonical.com>