feat(plugins): kernel: "relax" source requirements
The plugin has been relaxed to no longer explicitly require a specified
source for the part. However, we still need to have one. Here, we employ
a trick: if there is something which looks like a Linux kernel source
tree in CRAFT_PART_SRC, then we behave as a plugin normally would.
However, if there is nothing in our CRAFT_PART_SRC, see if the user is
building the snap in-tree. If so, treat CRAFT_PROJECT_DIR as
CRAFT_PROJECT_SRC.

The reason we do this is primarily to save on space and time -- it takes
a while even to copy the source from CRAFT_PROJECT_DIR to CRAFT_PART_SRC,
and it's even worse when cloning. This isn't an explicitly condoned
use-case as it violates standard plugin behavior, but the time savings
handling kernel sources in this way offers is worth the behavior-violating
trade-off.

This is made more palatable by the fact that this behavior isn't
necessarily destructive to the host environment anyways (only files in
debian*/ are changed), unless the user performs an override-pull or
something like it; in these cases, however, we can assume the user knows
what they are doing.

Signed-off-by: Dilyn Corner <dilyn.corner@canonical.com>